+++ src/main/java/webserver/Response.java
@@ -1,4 +1,4 @@
-package util;
+package webserver;
 
 import db.DataBase;
 import java.io.DataOutputStream;
